requirements.txt
numpy~=1.19.4
matplotlib~=3.3.2
Pillow~=8.0.1
opencv-python~=4.5.2.54